// src/app/core/models/facturation/expedition.model.ts
import { Document } from "./document.model";
import { ExpeditionItem } from "./expedition-item.model";
import { TarifDetail } from "./tarif-detail.model";

export class Expedition {
  id: number;
  adress_delivery: any;
  adress_expedition: any;
  agency_destination_id: number;
  agency_origin_id: number;
  amount: any;
  coming_back: boolean;
  comment: string;
  created_at: string;
  credit_type: string;
  current_hub_id: number;
  declaration_code: any;
  declaration_img: any;
  declared_value: string;
  delivery_date: any;
  delivery_matrices: string;
  delivery_matrix_index: number;
  delivery_matrix_log: any;
  demande_id: any;
  destination_city_id: any;
  editable: boolean;
  env_id: number[];
  expedition_items : ExpeditionItem[];
  ItemsProperties: { number: number, weight: number, declared_value: number };
  tarifs: { ht: number; tva: number; ttc: number; htTotal: number; ttcTotal: number };
  expedition_tarif_details : TarifDetail[];
  fully_returned: boolean;
  handed_to_agency: boolean;
  ht_amount: any;
  id_adress_expedition: any;
  id_adresse_delivery: any;
  id_customer: number;
  id_delivery: number;
  id_facture: number;
  id_invoice: any;
  id_receiver: number;
  id_service: number;
  import_hash: any;
  is_imported: boolean;
  is_overrided: boolean;
  is_taxed: boolean;
  mode_port: string;
  nature: string;
  on_travel: boolean;
  parent_id: any;
  pickup_date: any;
  pickup_houre: any;
  position_delivery: any;
  position_expedition: any;
  real_delivery_date: any;
  real_pickup_date: any;
  reference: string;
  return_documents:Document[];
  return_funds_amounts: string;
  return_funds_types: any;
  service_amount: any;
  status: string;
  step: number[];
  ticket_count: any;
  tour_id: any;
  transportation_amount: any;
  type_livraison: string;
  type_service: string;
  type_truck: any;
  updated_at: string;
  uuid: string;
  volume_truck: any;
  volumetric_weight: any;
}